U-6 Temporary Resolution
NOTE: There is currently a bug in Unity 6 that affects wheel colliders and center of gravity offset. This bug has been reported and is listed as a bug and is currently being worked on for an official fix. Until such we can provide a temporary resolution file to add to your vehicle to sort this out.